+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] <!-- ReleaseDate -->
 
 - add optional postcard MaxSize
+- add optional serde Serialize, Deserialize
 
 ## [2.2.0]  - 2023-08-24